lntest/itest: move single hop invoice test to separate file
